% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c_obj}
\alias{c_obj}
\title{Example COMPASS result object}
\format{
A COMPASSResult object with the following structure:
\describe{
  \item{fit}{Fitted model results from COMPASS analysis}
  \item{data}{Input data used for COMPASS analysis}
}
}
\usage{
c_obj
}
\description{
A COMPASSResult object containing example data for testing and
demonstration purposes.
}
\examples{
data("c_obj", package = "UtilsCompassSV")
class(c_obj)
}
\keyword{datasets}
