% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c_obj_list}
\alias{c_obj_list}
\title{Example list of COMPASS result objects}
\format{
A list with 4 elements, each a COMPASSResult object:
\describe{
  \item{Live Mtb}{Results for live Mycobacterium tuberculosis stimulation}
  \item{Secreted Mtb proteins}{Results for secreted Mtb protein stimulation}
  \item{Non-secreted Mtb proteins}{Results for non-secreted Mtb protein stimulation}
  \item{EBV/CMV}{Results for EBV/CMV stimulation}
}
}
\usage{
c_obj_list
}
\description{
A named list of COMPASSResult objects containing example data for
testing and demonstration purposes. Each element represents results
from different stimulation conditions.
}
\examples{
data("c_obj_list", package = "UtilsCompassSV")
names(c_obj_list)
}
\keyword{datasets}
