% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cyt_combn_format.R
\name{convert_cyt_combn_format}
\alias{convert_cyt_combn_format}
\title{Convert cytokine combination format to or from COMPASS format}
\usage{
convert_cyt_combn_format(
  cyt_combn,
  to,
  force = FALSE,
  silent = FALSE,
  check = TRUE
)
}
\arguments{
\item{cyt_combn}{character vector. Cytokine combination, specified either in
"<cyt><+-><cyt><+-><cyt><+->..." (e.g. IFNg+IL2-TNF+) or
COMPASS "<!><cyt>&<!><cyt>&<!><cyt>..." (e.g. IFNg&!IL2&TNF) format.}

\item{to}{'compass' or 'std'. Format to convert to. If \code{'compass'},
then output format is COMPASS format (specified above). If \code{'std'},
then output format is standard format (specified above).}

\item{force}{logical. If \code{TRUE}, then the code to convert to the specified format will be run.
If \code{FALSE}, then checks are made to ensure that conversion is not done if already in the
requested format. Default is \code{FALSE}.}

\item{silent}{logical. If \code{FALSE}, then warnings are printed if \code{force == FALSE} and
it is detected that \code{cyt_combn} already appears to be in the requested format. Default is \code{FALSE}.}

\item{check}{logical. If \code{TRUE}, then the converted format is checked that each element has
the same number of cytokines. Note that check is only performed if the conversion is attempted. Default is \code{TRUE}.}
}
\value{
character vector.
}
\description{
Convert cytokine combination format between COMPASS format and "standard " +/- format.
See \code{cyt_combn} parameter for details. NOTE: At present only converts from COMPASS to standard format.
}
\examples{
convert_cyt_combn_format(c("IFNg&!IL2"), to = 'std')

}
