% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_vec_grp.R
\name{.get_col_vec_grp}
\alias{.get_col_vec_grp}
\title{Get colours for groups for boxplots}
\usage{
.get_col_vec_grp(plot_prob_fill = NULL, .grp = NULL)
}
\arguments{
\item{plot_prob_fill}{character. If a colour, then the boxplots are filled according
to the specified value. If a named vector of colours names correspond to names
of the bottom-most level of the \code{c_obj} lists, then boxplots are filled
accordingly.  Otherwise, boxplots are filled according to names of lower-most list but using}

\item{.grp}{character vector. Names of groups in COMPASS object.}
}
\value{
A named character vector, where names are names of groups
and elements are colours for the groups
}
\description{
Get colours for groups for boxplots
}
\keyword{internal}
