% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compass-plot_compass_sub.R
\name{.plot_compass_pp}
\alias{.plot_compass_pp}
\title{Plot COMPASS posterior probabilites}
\usage{
.plot_compass_pp(
  c_obj,
  dir_save,
  prob_min,
  quant_min,
  silent,
  cyt_order,
  plot_prob_fill,
  facet,
  cyt_lab,
  boxplot_width,
  font_size,
  line_width,
  tile_colour,
  tile_alpha
)
}
\arguments{
\item{c_obj}{a named list of COMPASSResult objects. Provides data to plot.}

\item{dir_save}{character. Where to save the output. Default is working directory.}

\item{prob_min, quant_min}{[0,1]. Specify the minimum probability of a response for the minimum quantile
of samples that a cytokine combination must have to be included. For example,
if \code{prob_min == 0.5} and \code{quant_min == 0.1}, then only cytokine combinations for which
at least 10% of observations had a probability of a response of 0.5 or greater will plotted.
Default is \code{0.8} and \code{0.25}.}

\item{silent}{logical. If \code{TRUE}, then any warnings that would have been otherwise given
are not. Default is \code{FALSE}.}

\item{cyt_order}{character. If not \code{NULL}, then the order in which cytokines are arranged
vertically along the annotation grid is given by this vector (with first element going at the bottom).
If \code{NULL}, then cytokines are ordered by their order in COMPASS output. Default is \code{NULL}.}

\item{plot_prob_fill}{character. If a colour, then the boxplots are filled according
to the specified value. If a named vector of colours names correspond to names
of the bottom-most level of the \code{c_obj} lists, then boxplots are filled
accordingly.  Otherwise, boxplots are filled according to names of lower-most list but using}

\item{facet}{logical vector. Whether to facet or save individual plots for each
group in \code{c_obj}. If \code{TRUE} only, then only a faceted version is saved.
If \code{FALSE} only, then plots are saved individually. If \code{c(TRUE, FALSE)},
then both the facted and individual plots are saved.}

\item{cyt_lab}{function. A function that takes the cytokine names as input and returns the
value desired to be plotted along the y-axis of the grid plot. If not \code{NULL}, then
it is supplied when creating the grid plot to the \code{scale_y_continuous} function
via the \code{labels} parameter. For example, if we have two cytokines, \code{"IFNg"} and
\code{"TNF"}, but we want to display \code{"IFNg"} with the Greek gamma symbol, then we can
set \code{cyt_lab} equal to the following: \code{cyt_lab = function(cyt) purrr::map(cyt, function(cyt_ind){switch(cyt_ind, "IFNg" = bquote(paste(plain(paste("IFN")), gamma)), cyt_ind)})}.
This will change the label for \code{"IFNg"} but leave all the others as is.}

\item{tile_colour}{character. Colour to use for the tiles in the cytokine grid plot.
If \code{NULL}, then no colour is used. Default is \code{NULL}.}

\item{tile_alpha}{numeric in \code{[0,1]}. Transparency level to use for the tiles in the cytokine grid plot.
Default is \code{1} (no transparency, i.e. fully opaque).}
}
\value{
A named list with names 'p_probs' and 'p_grid',
corresponding to the list of ggplot2 plots of posterior plots
and the heat map of the corresponding cytokine combination
labels, respectively.
}
\description{
Plot COMPASS posterior probabilites
}
