% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_prob.R
\name{response_prob}
\alias{response_prob}
\title{Calculate COMPASS-derived overall responder probability}
\usage{
response_prob(c_obj)
}
\arguments{
\item{c_obj}{object of class 'COMPASSResult'. The posterior probabilities
for individual cytokine combinations are obtained here.}
}
\value{
A dataframe with columns sampleid and prob.
}
\description{
Calculate the probability of responding to at least one
cytokine combination, assuming independence of the probability estimates
between cytokine combinations.
}
\details{
Calculates the probability of an individual responding to
at least one cytokine combination as (1-product((1-prob_i))),
where prob_i is the probability of responding to the i-th cytokine
combination and the product is taken over all cytokine combinations
except the all-negative cytokine combination.
}
\examples{
data('c_obj', package = 'compassutils')
response_prob(c_obj = c_obj)

}
