% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compass-save_layout.R
\name{.save_plot}
\alias{.save_plot}
\title{Save an individual layout}
\usage{
.save_plot(p, height, width, dir_save, file, save_format, n_row, n_col)
}
\arguments{
\item{p}{object of class \code{gg}. Plot to save.}

\item{height}{numeric. Height of saved plot in centimeters.}

\item{width}{numeric. Width of saved plot in centimeters.}

\item{dir_save}{character. Where to save the output. Default is working directory.}

\item{file}{character. File name for the saved plot.}

\item{save_format}{"pdf" or "png". Plot device to use. Default is "png".}

\item{n_row, n_col}{integer. Number of objects plotted per row and column.
Determines \code{width} and \code{height} if they are \code{NULL}.}
}
\description{
Save an individual plot layout.
}
\keyword{internal}
