% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compass.R
\name{plot_compass}
\alias{plot_compass}
\title{Plot COMPASS output}
\usage{
plot_compass(
  c_obj,
  dir_save_base = getwd(),
  save = TRUE,
  save_format = "png",
  prob_min = 0.8,
  quant_min = 0.25,
  silent = FALSE,
  cyt_order = NULL,
  plot_prob_fill = NULL,
  plot_prob_shift = 0,
  return_plot = TRUE,
  facet = TRUE
)
}
\arguments{
\item{c_obj}{object of class 'COMPASSResult', or a list of such objects. Provides
COMPASS data to plot.}

\item{dir_save_base}{character. Where to save the output. Default is working directory.}

\item{save}{logical. If \code{TRUE}, then plots are saved. Default is \code{TRUE}.}

\item{save_format}{'pdf' or 'png'. Plot device to use. Default is 'png'.}

\item{prob_min, quant_min}{[0,1]. Specify the minimum probability of a response for the minimum quantile
of samples that a cytokine combination must have to be included. For example,
if \code{prob_min == 0.5} and quant_min == {0.1}, then only cytokine combinations for which
at least 10% of observations had a probability of a response of 0.5 or greater will plotted.
Default is \code{0.8} and \code{0.25}.}

\item{silent}{logical. If \code{TRUE}, then any warnings that would have been otherwise given
are not. Default is \code{FALSE}.}

\item{cyt_order}{character. If not \code{NULL}, then the order in which cytokines are arranged
vertically along the annotation grid is given by this vector (with first element going at the bottom).
If \code{NULL}, then cytokines are ordered by their order in COMPASS output. Default is \code{NULL}.}

\item{plot_prob_fill}{character. If a colour, then the boxplots are filled according
to the specified value. If a named vector of colours names correspond to names
of the bottom-most level of the \code{c_obj} lists, then boxplots are filled
accordingly.  Otherwise, boxplots are filled according to names of lower-most list but using}

\item{plot_prob_shift}{[0,1]. Extent to shift prob_plot. Increasing it from 0 moves the start of
the probability plot further to the right. Maximum value is 1 (at which point the plot will
effectively be pushed off the plotting surface). Useful to increase to a value such as 0.01
if the cytokine names are long and push the labelling grid too far to the right. Tweak as required.
Default is \code{0}.}

\item{return_plot}{logical. If \code{TRUE}, then plot output is returned.}

\item{facet}{logical vector. Whether to facet or save individual plots for each
group in \code{c_obj}. If \code{TRUE} only, then only a faceted version is saved.
If \code{FALSE} only, then plots are saved individually. If \code{c(TRUE, FALSE)},
then both the facted and individual plots are saved.}
}
\value{
A list, where each element is a \code{ggplot2} object.
}
\description{
Plot COMPASS output
}
\examples{
plot_compass(c_obj)

}
